完成二级回复界面，带隐藏

+++ src/main/java/com/sun/community/dto/CommentDTO.java
@@ -18,10 +18,9 @@ public class CommentDTO {
     private Long gmtCreate;
     private Long gmtModified;
     private Long likeCount;
-    private Integer commentCount;
+    private Long commentCount;
     private String content;
 
     private User user;
-
 
 }
